requirements.txt
factory-boy==2.11.1
mock==2.0.0
pyhamcrest==1.9.0
djangorestframework
django-cors-headers
django-river>=3.0.0
django-heroku
psycopg2
gunicorn
faker